+++ great_expectations/validator/validator.py
@@ -1,21 +1,7 @@
 """This is currently helping bridge APIs"""
 from great_expectations.dataset import PandasDataset, SqlAlchemyDataset, SparkDFDataset
+from great_expectations.dataset.sqlalchemy_dataset import SqlAlchemyBatchReference
 from great_expectations.util import load_class
-
-
-class SqlAlchemyTable(object):
-
-    def __init__(self, engine, table_name):
-        self._engine = engine
-        self._table_name = table_name
-
-    @property
-    def engine(self):
-        return self._engine
-
-    @property
-    def table_name(self):
-        return self._table_name
 
 
 class Validator(object):
@@ -46,16 +32,17 @@ class Validator(object):
             )
 
         elif issubclass(self.expectation_engine, SqlAlchemyDataset):
-            if not isinstance(self.batch.data, SqlAlchemyTable):
-                raise ValueError("SqlAlchemyDataset expectation_engine requires a SqlAlchemyTable for its batch")
+            if not isinstance(self.batch.data, SqlAlchemyBatchReference):
+                raise ValueError("SqlAlchemyDataset expectation_engine requires a SqlAlchemyBatchReference for its batch")
+            init_kwargs = self.batch.data.get_init_kwargs()
+            init_kwargs.update(self.init_kwargs)
             return self.expectation_engine(
-                table_name=self.batch.bdata.table_name,
-                engine=self.batch.data.engine,
                 batch_kwargs=self.batch.batch_kwargs,
-                batch_parameters=self.batc.bbatch_parameters,
-                batch_markers=self.batc.bbatch_markers,
-                data_context=self.batc.bdata_context,
-                expectation_suite=self.expectation_suite
+                batch_parameters=self.batch.batch_parameters,
+                batch_markers=self.batch.batch_markers,
+                data_context=self.batch.data_context,
+                expectation_suite=self.expectation_suite,
+                **init_kwargs
             )
 
         elif issubclass(self.expectation_engine, SparkDFDataset):